Define colors in lua config instead of TOML scheme
This commit moves terminal colors into the lua `colors` section and removes the
external TOML `color_scheme` files because of a change to WezTerm, which caused
tab bar colors in lua to be overridden by default colors when using a TOML color
scheme and the retro-style tab bar.

Previously, if both `color_scheme` and `colors` were specified in the lua
configuration, WezTerm used values from both the TOML file and the lua table
(with `color_scheme` winning if the same key was present in both). Starting in
a recent version, with the retro-style tab bar only, WezTerm now uses one or the
other but not both. If a `color_scheme` is specified, the retro-style tab bar
always appears in the default color scheme.

The fancy tab bar is not affected by the above change; if `use_fancy_tab_bar` is
set to `true`, WezTerm continues to use colors defined in both places.