Change layout to conform to WezTerm plugin system
Merge the theme files into a single Lua module, which gives access to
different variants of the theme.

WezTerm has a built-in plugin support since version 20230320-124340:
https://github.com/wez/wezterm/commit/e4ae8a844d8feaa43e1de34c5cc8b4f07ce525dd

A plugin is expected to have a single file at plugin/init.lua. Loading
multiple plugin lua files is not supported at the moment:
https://github.com/wez/wezterm/discussions/4150

Resolves #1